+++ src/ProjectMarworyn.Core/Generators/IPersonGenerator.cs
@@ -5,6 +5,6 @@ namespace ProjectMarworyn.Core.Generators
     internal interface IPersonGenerator
     {
         List<Person> Initialise(List<InitialPerson> initialPeople, int worldSeed);
-        (List<Person>, List<Person>) GenerateChildren(List<Pair> pairs, int worldSeed, int personId, List<Person> people, DateTime currentTime);
+        ChildGenerationResult GenerateChildren(List<Pair> pairs, int worldSeed, int personId, List<Person> people, DateTime currentTime);
     }
 }