+++ src/ProjectMarworyn.Core/Models/Person.cs
@@ -11,8 +11,9 @@ namespace ProjectMarworyn.Core.Models
         public Gender Gender { get; set; }
         public bool WillHaveChildren { get; set; }
         public bool IsAlive { get; set; }
-        public DateTime TimeLived { get; set; }
-        public DateTime TimeFromLastChild { get; set; }
+        public int BirthMonth { get; set; }
+        public int BirthDay { get; set; }
+        public int DaysSinceLastChild { get; set; }
         public bool HasPair { get; set; }
     }
 }